// src/modules/salary-grade/dependency-tier/interfaces/dependency-tier.interface.ts
export interface IDependencyTier {
  tierId: string;
  tierCode: string;
  description: string | null;

  attr1: string | null;
  attr2: string | null;
  attr3: string | null;
  attr4: string | null;
  attr5: string | null;

  isActive: string | null;
  isDeleted: boolean;

  createdBy: string;
  createdDate: string;

  modifiedBy: string | null;
  modifiedDate: string | null;
}

export interface IDependencyTierFilterOptions {
  search?: string;
  isActive?: string;
  page?: number;
  pageSize?: number;
}